+++ MemfaultBort/app/src/main/java/com/memfault/bort/AppComponents.kt
@@ -16,7 +16,7 @@ import kotlinx.serialization.json.JsonConfiguration
 import okhttp3.*
 import retrofit2.Converter
 import retrofit2.Retrofit
-import java.lang.IllegalArgumentException
+import java.util.*
 
 // A simple holder for application components
 // This example uses a service locator pattern over a DI framework for simplicity and ease of use;
@@ -27,25 +27,37 @@ data class AppComponents(
     val retrofitClient: Retrofit,
     val workerFactory: WorkerFactory,
     val fileUploaderFactory: FileUploaderFactory,
-    val bortEnabledProvider: BortEnabledProvider
+    val bortEnabledProvider: BortEnabledProvider,
+    val deviceIdProvider: DeviceIdProvider
 ) {
     open class Builder(
         private val context: Context,
-        private val sharedPreferences: SharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
+        private val sharedPreferences: SharedPreferences = PreferenceManager.getDefaultSharedPreferences(
+            context
+        )
     ) {
         var settingsProvider: SettingsProvider? = null
-        var networkInterceptor: Interceptor? = null
+        var loggingInterceptor: Interceptor? = null
+        var debugInfoInjectingInterceptor: Interceptor? = null
         var okHttpClient: OkHttpClient? = null
         var retrofit: Retrofit? = null
         var interceptingWorkerFactory: WorkerFactory? = null
         var fileUploaderFactory: FileUploaderFactory? = null
         var bortEnabledProvider: BortEnabledProvider? = null
+        var deviceIdProvider: DeviceIdProvider? = null
 
         fun build(): AppComponents {
             val settingsProvider = settingsProvider ?: BuildConfigSettingsProvider()
+            val deviceIdProvider = deviceIdProvider ?: RandomUuidDeviceIdProvider(sharedPreferences)
             val fileUploaderFactory = fileUploaderFactory ?: MemfaultFileUploaderFactory()
             val okHttpClient = okHttpClient ?: OkHttpClient.Builder()
-                .addInterceptor(networkInterceptor ?: LoggingNetworkInterceptor())
+                .addInterceptor(
+                    debugInfoInjectingInterceptor ?: DebugInfoInjectingInterceptor(
+                        settingsProvider,
+                        deviceIdProvider
+                    )
+                )
+                .addInterceptor(loggingInterceptor ?: LoggingNetworkInterceptor())
                 .build()
             val retrofit = retrofit ?: Retrofit.Builder()
                 .client(okHttpClient)
@@ -54,14 +66,15 @@ data class AppComponents(
                     kotlinxJsonConverterFactory()
                 )
                 .build()
-            val bortEnabledProvider = bortEnabledProvider ?: if (settingsProvider.isRuntimeEnableRequired()) {
-                PreferenceBortEnabledProvider(
-                    sharedPreferences,
-                    defaultValue = !settingsProvider.isRuntimeEnableRequired()
-                )
-            } else {
-                BortAlwaysEnabledProvider()
-            }
+            val bortEnabledProvider =
+                bortEnabledProvider ?: if (settingsProvider.isRuntimeEnableRequired()) {
+                    PreferenceBortEnabledProvider(
+                        sharedPreferences,
+                        defaultValue = !settingsProvider.isRuntimeEnableRequired()
+                    )
+                } else {
+                    BortAlwaysEnabledProvider()
+                }
 
             val workerFactory = DefaultWorkerFactory(
                 settingsProvider = settingsProvider,
@@ -71,12 +84,13 @@ data class AppComponents(
                 interceptingFactory = interceptingWorkerFactory
             )
             return AppComponents(
-                settingsProvider =  settingsProvider,
+                settingsProvider = settingsProvider,
                 okHttpClient = okHttpClient,
                 retrofitClient = retrofit,
                 workerFactory = workerFactory,
                 fileUploaderFactory = fileUploaderFactory,
-                bortEnabledProvider = bortEnabledProvider
+                bortEnabledProvider = bortEnabledProvider,
+                deviceIdProvider = deviceIdProvider
             )
         }
     }
@@ -88,7 +102,7 @@ fun kotlinxJsonConverterFactory(): Converter.Factory =
     Json(JsonConfiguration.Stable)
         .asConverterFactory(MediaType.get("application/json"))
 
-class LoggingNetworkInterceptor: Interceptor {
+class LoggingNetworkInterceptor : Interceptor {
     override fun intercept(chain: Interceptor.Chain): Response {
         val request: Request = chain.request()
         val t1: Long = System.nanoTime()
@@ -105,7 +119,74 @@ Received response for ${response.request().url()} in ${String.format("%.1f", del
     }
 }
 
-open class BuildConfigSettingsProvider: SettingsProvider {
+internal const val QUERY_PARAM_UPSTREAM_VERSION_NAME = "upstreamVersionName"
+internal const val QUERY_PARAM_UPSTREAM_VERSION_CODE = "upstreamVersionCode"
+internal const val QUERY_PARAM_UPSTREAM_GIT_SHA = "upstreamGitSha"
+internal const val QUERY_PARAM_VERSION_NAME = "versionName"
+internal const val QUERY_PARAM_VERSION_CODE = "versionCode"
+internal const val QUERY_PARAM_DEVICE_ID = "deviceId"
+internal const val X_REQUEST_ID = "X-Request-Id"
+
+/** Inject debug information for requests to memfault or to a local api server */
+class DebugInfoInjectingInterceptor(
+    private val settingsProvider: SettingsProvider,
+    private val deviceIdProvider: DeviceIdProvider
+) : Interceptor {
+
+    fun transformRequest(request: Request): Request {
+        with (request.url()) {
+            when {
+                this.host() in listOf("127.0.0.1", "localhost")
+                        && request.url().encodedPathSegments().firstOrNull() == "api" -> {
+                    // Fall through
+                }
+                this.topPrivateDomain().equals("memfault.com", ignoreCase = true) -> {
+                    // Fall through
+                }
+                else -> return request
+            }
+        }
+        return request.url().newBuilder().apply {
+            listOf(
+                QUERY_PARAM_UPSTREAM_VERSION_NAME to settingsProvider.upstreamVersionName(),
+                QUERY_PARAM_UPSTREAM_VERSION_CODE to settingsProvider.upstreamVersionCode()
+                    .toString(),
+                QUERY_PARAM_UPSTREAM_GIT_SHA to settingsProvider.upstreamGitSha(),
+                QUERY_PARAM_VERSION_NAME to settingsProvider.appVersionName(),
+                QUERY_PARAM_VERSION_CODE to settingsProvider.appVersionCode().toString(),
+                QUERY_PARAM_DEVICE_ID to deviceIdProvider.deviceId()
+            ).forEach {
+                addQueryParameter(it.first, it.second)
+            }
+        }.let {
+            request.newBuilder()
+                .addHeader(X_REQUEST_ID, UUID.randomUUID().toString())
+                .url(it.build())
+                .build()
+        }
+    }
+
+    override fun intercept(chain: Interceptor.Chain): Response =
+        chain.proceed(
+            transformRequest(chain.request())
+        )
+}
+
+interface DeviceIdProvider {
+    fun deviceId(): String
+}
+
+class RandomUuidDeviceIdProvider(
+    sharedPreferences: SharedPreferences
+) : DeviceIdProvider, PreferenceKeyProvider<String>(
+    sharedPreferences = sharedPreferences,
+    defaultValue = UUID.randomUUID().toString(),
+    preferenceKey = PREFERENCE_DEVICE_ID
+) {
+    override fun deviceId(): String = super.getValue()
+}
+
+open class BuildConfigSettingsProvider : SettingsProvider {
     override fun bugReportRequestIntervalHours() =
         BuildConfig.BUG_REPORT_REQUEST_INTERVAL_HOURS.toLong()
 
@@ -126,6 +207,18 @@ open class BuildConfigSettingsProvider: SettingsProvider {
     override fun projectKey(): String = BuildConfig.MEMFAULT_PROJECT_API_KEY
 
     override fun baseUrl(): String = BuildConfig.MEMFAULT_FILES_BASE_URL
+
+    override fun appVersionName(): String = BuildConfig.VERSION_NAME
+
+    override fun appVersionCode(): Int = BuildConfig.VERSION_CODE
+
+    override fun upstreamVersionName(): String = BuildConfig.UPSTREAM_VERSION_NAME
+
+    override fun upstreamVersionCode(): Int = BuildConfig.UPSTREAM_VERSION_CODE
+
+    override fun upstreamGitSha(): String = BuildConfig.UPSTREAM_GIT_SHA
+
+    override fun currentGitSha(): String = BuildConfig.CURRENT_GIT_SHA
 }
 
 class DefaultWorkerFactory(
@@ -191,13 +284,14 @@ abstract class PreferenceKeyProvider<T>(
 ) {
     init {
         when (defaultValue) {
-            is Boolean, is String, is Int, is Long, is Float -> {}
+            is Boolean, is String, is Int, is Long, is Float -> {
+            }
             else -> throw IllegalArgumentException("Unsupported type $defaultValue")
         }
     }
 
     @Suppress("UNCHECKED_CAST")
-    fun setValue(newValue : T): Unit = with(sharedPreferences.edit()) {
+    fun setValue(newValue: T): Unit = with(sharedPreferences.edit()) {
         when (newValue) {
             is Boolean -> putBoolean(preferenceKey, newValue)
             is String -> putString(preferenceKey, newValue)