+++ MemfaultBort/app/src/main/java/com/memfault/bort/Constants.kt
@@ -9,4 +9,6 @@ const val APPLICATION_ID_MEMFAULT_USAGE_REPORTER = "com.memfault.usagereporter"
 
 const val INTENT_ACTION_BORT_ENABLE = "com.memfault.intent.action.BORT_ENABLE"
 const val INTENT_EXTRA_BORT_ENABLED = "com.memfault.intent.extra.BORT_ENABLED"
+
 const val PREFERENCE_BORT_ENABLED = "com.memfault.preference.BORT_ENABLED"
+const val PREFERENCE_DEVICE_ID = "com.memfault.preference.DEVICE_ID"