+++ MemfaultBort/app/src/main/java/com/memfault/bort/receivers/BortEnableReceiver.kt
@@ -16,6 +16,7 @@ class BortEnableReceiver : SingleActionReceiver(
         )
         val wasEnabled = bortEnabledProvider.isEnabled()
         Logger.test("wasEnabled=$wasEnabled isNowEnabled=$isNowEnabled")
+        Logger.logEventBortSdkEnabled(isNowEnabled)
         if (wasEnabled == isNowEnabled) {
             return
         }