+++ MemfaultBort/app/src/main/java/com/memfault/bort/receivers/BugReportReceiver.kt
@@ -2,6 +2,7 @@ package com.memfault.bort.receivers
 
 import android.content.Context
 import android.content.Intent
+import android.util.EventLog
 import com.memfault.bort.*
 import com.memfault.bort.uploader.BugReportUploadScheduler
 import java.io.File
@@ -12,6 +13,7 @@ class BugReportReceiver : BortEnabledFilteringReceiver(
     override fun onReceivedAndEnabled(context: Context, intent: Intent) {
         val bugreportPath = intent.getStringExtra(INTENT_EXTRA_BUGREPORT_PATH)
         Logger.v("Got bugreport path: $bugreportPath")
+        Logger.logEvent("bugreport", "received")
         bugreportPath ?: return
 
         val bugreportFile = File(bugreportPath)